test(sessions): pin detach memo invalidation to the equal-size rewrite
